fix: hook을 분리해보앗어용
